requirements.txt
pandas==1.5.3
requests==2.28.2
sqlalchemy==1.4.46
prophet==1.1.2
python-dotenv==1.0.0
numpy==1.23.5
psycopg2
holidays==0.10.3
plotly